refactor: reorganize settings — Windows pane + Contents → Behavior
Two information-architecture cleanups:

- Window management moves out of the Shortcuts pane into a new dedicated
  "Windows" pane, split into "Arrange window" (tile/maximize/center/restore +
  cycle-tile-widths) and "All windows" (hide-all/show-all + keep-apps-visible).
  Shortcuts now focuses on the switcher triggers, direct activation, and the
  in-panel keys.
- The "Contents" options (which windows/apps the switcher lists: show
  minimized/hidden/windowless, applications-only, badges, current Space, sort
  order, recently closed) move from the Appearance pane into the Behavior pane,
  since they decide *which* windows show — behavior, not look. Appearance is now
  purely visual (layout, size, grid, accent, opacity, corner radius, title).

Search catalog and section anchors updated to match; the moved search items now
navigate to their new panes.
