+++ BetterCmdTab/Settings/ShortcutsSettingsViewController.swift
@@ -13,14 +13,6 @@ final class ShortcutsSettingsViewController: SettingsTabViewController {
     private var directButtons: [NSButton] = []
     private var directSlotSheet: AppsPickerSheetWindowController?
 
-    // Window-management options.
-    private let cycleWidthsSwitch = NSSwitch()
-
-    // "Hide all windows" exclusion list: a row whose subtitle shows the count
-    // and a picker sheet to edit which apps stay visible.
-    private var excludedHideAppsRow: SettingsRowView?
-    private var excludedHideAppsSheet: AppsPickerSheetWindowController?
-
     override func setupContent() {
         // Switcher shortcuts — the unified, AltTab-style tabbed editor: the two
         // core triggers (Apps, Windows) plus each user-created scoped shortcut,
@@ -70,55 +62,6 @@ final class ShortcutsSettingsViewController: SettingsTabViewController {
         for (name, title) in BetterShortcuts.Name.panelActionKeys {
             addRow(to: panelKeys, title: title, accessory: BetterShortcuts.RecorderCocoa(for: name))
         }
-
-        // Window management section — tile / maximize / center. These ARE global
-        // shortcuts (they work whether the switcher is open or closed); when the
-        // they always arrange the frontmost app's focused window, whether or
-        // not the switcher is open. Default ⌃⌘ + arrows.
-        let windowMgmt = addSection(title: String(localized: "Window management"), anchor: SettingsAnchor.windowMgmt)
-        addRow(
-            to: windowMgmt,
-            title: String(localized: "Arrange the focused window"),
-            subtitle: String(localized: "Tile to a half or corner, maximize, or center the frontmost window. Works system-wide."),
-            searchItemID: SearchID.windowMgmt
-        )
-        for (name, title) in BetterShortcuts.Name.windowMgmt {
-            addRow(to: windowMgmt, title: title, accessory: BetterShortcuts.RecorderCocoa(for: name))
-        }
-        addRow(
-            to: windowMgmt,
-            title: String(localized: "Hide all windows"),
-            subtitle: String(localized: "Hide every app to reveal the desktop. Works system-wide."),
-            accessory: BetterShortcuts.RecorderCocoa(for: .hideAllWindows)
-        )
-        addRow(
-            to: windowMgmt,
-            title: String(localized: "Show all windows"),
-            subtitle: String(localized: "Bring every hidden app back."),
-            accessory: BetterShortcuts.RecorderCocoa(for: .showAllWindows)
-        )
-        let excludeButton = NSButton(
-            title: String(localized: "Choose…"),
-            target: self,
-            action: #selector(chooseExcludedHideApps)
-        )
-        excludeButton.bezelStyle = .rounded
-        excludeButton.controlSize = .small
-        excludedHideAppsRow = addRow(
-            to: windowMgmt,
-            title: String(localized: "Keep apps visible"),
-            subtitle: Self.excludedHideDescription(Preferences.shared.hideAllExcludedBundleIDs.count),
-            accessory: excludeButton
-        )
-        cycleWidthsSwitch.controlSize = .small
-        cycleWidthsSwitch.target = self
-        cycleWidthsSwitch.action = #selector(toggleCycleWidths(_:))
-        addRow(
-            to: windowMgmt,
-            title: String(localized: "Cycle tile widths"),
-            subtitle: String(localized: "Press Tile left / Tile right again to step the window through ½ → ⅔ → ⅓ of the screen on that side."),
-            accessory: cycleWidthsSwitch
-        )
     }
 
     override func viewWillAppear() {
@@ -127,47 +70,6 @@ final class ShortcutsSettingsViewController: SettingsTabViewController {
         // Another pane (Import settings) can rewrite the shortcut list/overrides
         // off-screen; rebuild the editor from the live model on appear.
         shortcutsEditorView.reload()
-        cycleWidthsSwitch.state = Preferences.shared.cycleTileWidths ? .on : .off
-        // Another pane (e.g. Import settings) can rewrite the list while this
-        // cached controller is off screen — re-sync the subtitle on appear.
-        excludedHideAppsRow?.update(
-            subtitle: Self.excludedHideDescription(Preferences.shared.hideAllExcludedBundleIDs.count)
-        )
-    }
-
-    @objc private func toggleCycleWidths(_ sender: NSSwitch) {
-        Preferences.shared.cycleTileWidths = (sender.state == .on)
-    }
-
-    /// Subtitle for the "Keep apps visible" row: explains the empty state, else
-    /// reports how many apps are excluded from Hide all windows.
-    private static func excludedHideDescription(_ count: Int) -> String {
-        if count == 0 {
-            return String(localized: "Hide all windows hides every app, Finder included. Pick apps to keep visible.")
-        }
-        return String(localized: "Apps kept visible: \(count).")
-    }
-
-    /// Open the multi-select app picker seeded with the current exclusions; the
-    /// returned set replaces the stored list. The picker itself is the whole
-    /// management UI — no per-row remove needed.
-    @objc private func chooseExcludedHideApps() {
-        guard let window = view.window, excludedHideAppsSheet == nil else { return }
-        let current = Set(Preferences.shared.hideAllExcludedBundleIDs)
-        let controller = AppsPickerSheetWindowController(
-            title: String(localized: "Keep apps visible"),
-            prompt: String(localized: "Chosen apps stay visible when you trigger Hide all windows."),
-            selectedBundleIDs: current,
-            singleSelection: false,
-            confirmTitle: String(localized: "Done")
-        ) { [weak self] selection in
-            guard let self else { return }
-            Preferences.shared.hideAllExcludedBundleIDs = selection.sorted()
-            self.excludedHideAppsRow?.update(subtitle: Self.excludedHideDescription(selection.count))
-        }
-        controller.onDidDismiss = { [weak self] in self?.excludedHideAppsSheet = nil }
-        excludedHideAppsSheet = controller
-        controller.present(asSheetFor: window)
     }
 
     // MARK: - Direct activation slots
